serverChatBot/requirements.txt
transformers
langchain
langchain-huggingface
pinecone
langchain-core 
langgraph>0.2.27
-qU "langchain[groq]"
torch
request
accelerate